http: Fixed null values and date info in output of Anilist API query
Fixed some bugs in the end date column, pre-emptively dealt with null
values. Currently contemplating making the goal predicting the number of
chapters, since I was planning on using random forest regression, and
that is ill-suited to predicting values outside of the range of the
training set, which we would need to do if we were planning to predict
future dates for unfinished manga.