http: changed json parsing method, handling of missing dates
Changed handling of dates in data gathering script to simply record what
data was given, with the intent of giving more attention to the method
for handling missingness later on during data wrangling in jupyter
notebook.

Additionally, changed the JSON parsing loop to initialize the dictionary
in a more readable way.