http: streamlined json parsing code
Rewrote JSON parsing code to use defultdict instead of using conditions
to check if a key existed and then initializing it repeatedly. This
results in much cleaner code, although it means we will have to handle
data wrangling differently, as NAs will be filled with zero in factor
variables as well.

Simplified relation and character parsing by using procedural naming.