gather/experiment/dir: reorganized project
Project will now be broken up into phases in order to make its structure
clearer, allow for individual components to be completed without
including in-progress code for future components.

Changed readme for overarching project to not refer to the project as a
single script.

Created folders for each segment of the project:
-data gathering refers to the project that calls the AniList API and
saves the response as a csv file.
-data exploration will involve visualizing and calculating metrics for
the data we gathered in order to better understand its structure.
-experiments is where the work I have completed that does not fit into
either of these structures will go for now.

In the experiments project, I began the process of imputing chapter
values on their own, so that all chapter and volume values would be
imputed by the process of using the ratio to impute chapters from
volumes and vice-versa. The goal was to use random forest regression,
utilizing run length as a predictor where possible, and running a
separate model without run length as a predictor for cases where it did
not exist.

I am beginning to realize that the methods I am using for data
imputation might be leading me in the wrong direction, and I intend to
revisit this later.