readme/data/gather: cleanup of data dictionary, related
readme:
Fixed some links to portions of the project so that they work properly.

data:
Improved clarity around the names of dynamically named features.

gather:
Removed the method of prepending tag_ to dynamically named tag features.
This was initially intended to keep track of the features and refer to
them collectively more easily, but this can more cleanly be done by
utilizing the appendix to the data dictionary to get a full list of
names and categories for each tag.