data_dictionary.md: Improve markdown rendering of feature definitions
Pre-patch, each feature was formatted as a new h3 header (`###`) and
subsequent paragraph. Ideally, this file would use the markdown syntax
for definition lists[1]. However, the syntax is not universally
supported, namely not by GitHub Flavored Markdown[2].  Therefore, this
patch replaces each h3 header and subsequent paragraph with a single
paragraph, where the first word (the feature name) is bold and followed
by a hard linebreak.

In GitHub's web ui, this significantly improves the readability.

[1]: https://www.markdownguide.org/extended-syntax/#definition-lists
[2]: https://github.github.com/gfm