scripts: Fix when run from unexpected dir
Pre-patch the project's scripts wrote output to '../data'. If the
script's cwd was not the script's parent dir, it would cause script
failure or cause output written to the wrong directory

Allow running the scripts to be run from any directory, such as the
project top dir. The scripts write their output to the same intended
location as before, '{project_top}/data/'.